Api_QA/requirements.txt
streamlit
requests
pandas
pyyaml
bcrypt
google-generativeai

